requirements.txt
Pillow
tkinter
```